Moved select search option to search directory
