+++ client/src/components/rides/sections/DetailViewSection.tsx
@@ -2,7 +2,7 @@ import { Box, Stack, TextField, Typography } from "@mui/material";
 import { SelectChangeEvent } from "@mui/material/Select";
 import { useState } from "react";
 import { getRideData, RideData } from "../../../data/rideData";
-import { SelectSearchOption } from "../detail-view/SelectSearchOption";
+import { SelectSearchOption } from "../detail-view/search/SelectSearchOption";
 import { getRequestTimeString } from "../../../utils/utils";
 import { RideDataFullDetails } from "../detail-view/RideDataFullDetails";
 import { AccordionWrapper } from "../detail-view/AccordionWrapper";
